# Proposal of a predicate function