My proposal for improving situations like #25