@oflebbe 2021-04-02
benchmark                       old ns/op     new ns/op     delta
BenchmarkLondon-8               278851746     274827393     -1.44%
BenchmarkLondon_nodes-8         181481979     189889038     +4.63%
BenchmarkLondon_ways-8          161872523     168744234     +4.25%
BenchmarkLondon_relations-8     98999793      102736629     +3.77%

benchmark                       old allocs     new allocs     delta
BenchmarkLondon-8               2416803        2416809        +0.00%
BenchmarkLondon_nodes-8         1003854        1003848        -0.00%
BenchmarkLondon_ways-8          1792716        1792715        -0.00%
BenchmarkLondon_relations-8     456774         456773         -0.00%

benchmark                       old bytes     new bytes     delta
BenchmarkLondon-8               954510768     954510898     +0.00%
BenchmarkLondon_nodes-8         649114912     649114142     -0.00%
BenchmarkLondon_ways-8          432451373     432451284     -0.00%
BenchmarkLondon_relations-8     178718509     178718332     -0.00%