@paulmach 2022-02-08
I finished this up over here https://github.com/paulmach/osm/pull/30

It's kind of a "visit" thing. Faster if you're okay with the filter functions being calling in parallel and not if a predefined order.